requirements.txt
sec-api-python
scikit-learn
yfinance
pandas
numpy
